Refactor view model messaging architecture and base classes
Introduce AppMessagingViewModelBase and DocumentMessagingViewModelBase to clearly separate application-scoped and document-scoped messaging. Remove DocumentViewModelBase in favor of the new modular base classes. Update ViewModelBase to inherit from ObservableRecipient and add constructors for messenger injection. Update MermaidEditorViewModel to use the new base class and clarify messaging documentation. Add MIT license headers and improve XML docs throughout. This refactor improves maintainability, testability, and message subscription safety.
